eghegheghegh

+++ lib/ravLidar.cpp
@@ -104,7 +104,7 @@ int listenLidar(){
             int quality = (int)(buffer[0])>>2;
             int angle = (((int)(buffer[1])>>1) + ((int)(buffer[2])<<8))>>7;
             float theta = angle * (PI / 180);
-            float distance = (((float)(buffer[3])) + ((float)(buffer[4])<<8))>>2;
+            float distance = (float)((((int)(buffer[3])) + ((int)(buffer[4])<<8))>>2);
             //printf("radius: %d, angle: %d\n", distance, angle);
 
             points.row(iter) << distance, theta;
